Adapt the tests to Scala 3.5
