+++ tests/neg/i6716.check
@@ -1,5 +1,5 @@
--- Warning: tests/neg/i6716.scala:12:39 --------------------------------------------------------------------------------
-12 |  given Monad[Bar] = summon[Monad[Foo]] // warn
+-- Error: tests/neg/i6716.scala:11:39 ----------------------------------------------------------------------------------
+11 |  given Monad[Bar] = summon[Monad[Foo]] // error
    |                                       ^
    |                             Result of implicit search for Monad[Foo] will change.
    |                             Current result Bar.given_Monad_Bar will be no longer eligible
@@ -12,5 +12,3 @@
    |                               - use a `given ... with` clause as the enclosing given,
    |                               - rearrange definitions so that Bar.given_Monad_Bar comes earlier,
    |                               - use an explicit argument.
-   |                             This will be an error in Scala 3.5 and later.
-No warnings can be incurred under -Werror (or -Xfatal-warnings)
