+++ tests/neg/i6716.scala
@@ -1,4 +1,3 @@
-//> using options -Xfatal-warnings
 
 trait Monad[T]:
   def id: String
@@ -9,11 +8,10 @@ object Foo {
 
 opaque type Bar = Foo
 object Bar {
-  given Monad[Bar] = summon[Monad[Foo]] // warn
+  given Monad[Bar] = summon[Monad[Foo]] // error
 }
 
 object Test extends App {
   println(summon[Monad[Foo]].id)
   println(summon[Monad[Bar]].id)
 }
-// nopos-error: No warnings can be incurred under -Werror (or -Xfatal-warnings)
