+++ tests/neg/looping-givens.check
@@ -1,0 +1,48 @@
+-- Error: tests/neg/looping-givens.scala:9:22 --------------------------------------------------------------------------
+9 |  given aa: A = summon       // error
+  |                      ^
+  |                      Result of implicit search for T will change.
+  |                      Current result ab will be no longer eligible
+  |                        because it is not defined before the search position.
+  |                      Result with new rules: a.
+  |                      To opt into the new rules, compile with `-source future` or use
+  |                      the `scala.language.future` language import.
+  |
+  |                      To fix the problem without the language import, you could try one of the following:
+  |                        - use a `given ... with` clause as the enclosing given,
+  |                        - rearrange definitions so that ab comes earlier,
+  |                        - use an explicit argument.
+  |
+  |                      where:    T is a type variable with constraint <: A
+-- Error: tests/neg/looping-givens.scala:10:22 -------------------------------------------------------------------------
+10 |  given bb: B = summon       // error
+   |                      ^
+   |                      Result of implicit search for T will change.
+   |                      Current result ab will be no longer eligible
+   |                        because it is not defined before the search position.
+   |                      Result with new rules: b.
+   |                      To opt into the new rules, compile with `-source future` or use
+   |                      the `scala.language.future` language import.
+   |
+   |                      To fix the problem without the language import, you could try one of the following:
+   |                        - use a `given ... with` clause as the enclosing given,
+   |                        - rearrange definitions so that ab comes earlier,
+   |                        - use an explicit argument.
+   |
+   |                      where:    T is a type variable with constraint <: B
+-- Error: tests/neg/looping-givens.scala:11:28 -------------------------------------------------------------------------
+11 |  given ab: (A & B) = summon // error
+   |                            ^
+   |                            Result of implicit search for T will change.
+   |                            Current result ab will be no longer eligible
+   |                              because it is not defined before the search position.
+   |                            Result with new rules: Search Failure: joint(ab, ab).
+   |                            To opt into the new rules, compile with `-source future` or use
+   |                            the `scala.language.future` language import.
+   |
+   |                            To fix the problem without the language import, you could try one of the following:
+   |                              - use a `given ... with` clause as the enclosing given,
+   |                              - rearrange definitions so that ab comes earlier,
+   |                              - use an explicit argument.
+   |
+   |                            where:    T is a type variable with constraint <: A & B
