+++ tests/run/t2029.scala
@@ -5,7 +5,7 @@ object Test{
     val mainSet = TreeSet(1 to 5 :_*)
 
     var compareCalled = false;
-    val smallerSet = TreeSet(2 to 4 :_*)(Ordering[Int].reverse)
+    val smallerSet = TreeSet(2 to 4 :_*)(using Ordering[Int].reverse)
 
     println(mainSet.mkString(","))
     println(smallerSet.mkString(","))
